reorganize files, use socket operations
