+++ pkg/networkpath/traceroute/tcp/utils_unix_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build test
+//go:build unix && test
 
 package tcp
 
@@ -52,14 +52,15 @@ type (
 
 func Test_reserveLocalPort(t *testing.T) {
 	// WHEN we reserve a local port
-	port, listener, err := reserveLocalPort()
+	port, listener, err := reserveLocalPort(0)
 	require.NoError(t, err)
 	require.NotNil(t, listener)
+	assert.Equal(t, 161, int(port))
 
 	// THEN we should not be able to get another connection
 	// on the same port
 	conn2, err := net.Listen("tcp", fmt.Sprintf("0.0.0.0:%d", port))
-	assert.Error(t, err)
+	require.Error(t, err)
 	assert.Contains(t, err.Error(), "address already in use")
 	assert.Nil(t, conn2)
 }
